Move the Collection trait into its own module
